+++ actionpack/lib/action_dispatch/routing/route_set.rb
@@ -130,7 +130,7 @@ module ActionDispatch
             @url_helpers_module.send  :undef_method, url_name
           end
           routes[key] = route
-          define_url_helper @path_helpers_module, route, path_name, route.defaults, name, LEGACY
+          define_url_helper @path_helpers_module, route, path_name, route.defaults, name, PATH
           define_url_helper @url_helpers_module,  route, url_name,  route.defaults, name, UNKNOWN
 
           @path_helpers << path_name
@@ -319,34 +319,7 @@ module ActionDispatch
       # :stopdoc:
       # strategy for building urls to send to the client
       PATH    = ->(options) { ActionDispatch::Http::URL.path_for(options) }
-      FULL    = ->(options) { ActionDispatch::Http::URL.full_url_for(options) }
       UNKNOWN = ->(options) { ActionDispatch::Http::URL.url_for(options) }
-      LEGACY  = ->(options) {
-        if options.key?(:only_path)
-          if options[:only_path]
-            ActiveSupport::Deprecation.warn(<<-MSG.squish)
-              You are calling a `*_path` helper with the `only_path` option
-              explicitly set to `true`. This option will stop working on
-              path helpers in Rails 5. Simply remove the `only_path: true`
-              argument from your call as it is redundant when applied to a
-              path helper.
-            MSG
-
-            PATH.call(options)
-          else
-            ActiveSupport::Deprecation.warn(<<-MSG.squish)
-              You are calling a `*_path` helper with the `only_path` option
-              explicitly set to `false`. This option will stop working on
-              path helpers in Rails 5. Use the corresponding `*_url` helper
-              instead.
-            MSG
-
-            FULL.call(options)
-          end
-        else
-          PATH.call(options)
-        end
-      }
       # :startdoc:
 
       attr_accessor :formatter, :set, :named_routes, :default_scope, :router